NO JIRA ISSUE
The zip files received from the client where deleted as soon as the combined
zip was unpacked. The disadvantage of this was that is was then impossible to
examine the zip file if the bag was found to be invalid. Examining the zip file
could be useful to trace back the problem to the client (or determine that the
the error was caused on the server).

I have now moved the step to delete the zip files to after the validation. It will
only be done if the bag is found to be VALID.